+++ examples/2D_BFS_highRe_kwSST.jl
@@ -1,0 +1,149 @@
+using Plots
+using XCALibre
+# using CUDA
+
+# mesh_file = "unv_sample_meshes/backwardFacingStep_5mm.unv"
+# backwardFacingStep_2mm, 5mm or 10mm
+grids_dir = pkgdir(XCALibre, "examples/0_GRIDS")
+grid = "backwardFacingStep_10mm.unv"
+mesh_file = joinpath(grids_dir, grid)
+
+mesh = UNV2D_mesh(mesh_file, scale=0.001)
+
+# backend = CUDABackend(); workgroup = 32
+backend = CPU(); workgroup = 1024; activate_multithread(backend)
+
+hardware = Hardware(backend=backend, workgroup=workgroup)
+mesh_dev = adapt(backend, mesh)
+
+nu = 1e-3
+u_mag = 1.5 # 5mm mesh
+# u_mag = 3.5 # 2mm mesh
+velocity = [69.4, 0.0, 0.0]
+Tu = 0.05
+nuR = 100
+k_inlet = 1 #3/2*(Tu*u_mag)^2
+ω_inlet = 1000 #k_inlet/(nuR*nu)
+νt_inlet = k_inlet/ω_inlet
+Re = velocity[1]*0.1/nu
+
+model = Physics(
+    time = Steady(),
+    fluid = Fluid{Incompressible}(nu = nu),
+    turbulence = RANS{KOmegaSST}(walls=(:wall,)),
+    energy = Energy{Isothermal}(),
+    domain = mesh_dev
+    )
+
+BCs = assign(
+    region = mesh_dev,
+    (
+        U = [
+            Dirichlet(:inlet, velocity),
+            Zerogradient(:outlet),
+            Wall(:wall, [0.0, 0.0, 0.0]),
+            Wall(:top, [0.0, 0.0, 0.0])
+        ],
+        p = [
+            Neumann(:inlet, 0.0),
+            Dirichlet(:outlet, 0.0),
+            Wall(:wall),
+            Wall(:top)
+        ],
+        k = [
+            Dirichlet(:inlet, k_inlet),
+            Zerogradient(:outlet),
+            KWallFunction(:wall),
+            KWallFunction(:top)
+        ],
+        omega = [
+            Dirichlet(:inlet, ω_inlet),
+            Zerogradient(:outlet),
+            OmegaWallFunction(:wall),
+            OmegaWallFunction(:top)
+        ],
+        nut = [
+            Dirichlet(:inlet, νt_inlet),
+            Zerogradient(:outlet),
+            NutWallFunction(:wall), 
+            NutWallFunction(:top)
+        ]
+    )
+)
+
+schemes = (
+    U = Schemes(divergence=Upwind),
+    p = Schemes(divergence=Upwind),
+    k = Schemes(divergence=Upwind),
+    omega = Schemes(divergence=Upwind),
+    y = Schemes(gradient=Midpoint)
+)
+
+solvers = (
+    U = SolverSetup(
+        solver      = Bicgstab(), # Bicgstab(), Gmres()
+        preconditioner = Jacobi(),
+        convergence = 1e-7,
+        relax       = 0.7,
+        rtol = 1e-2,
+        atol = 1e-10
+    ),
+    p = SolverSetup(
+        solver      = Cg(), #Gmres(), #Cg(), # Bicgstab(), Gmres()
+        preconditioner = Jacobi(),
+        convergence = 1e-7,
+        relax       = 0.3,
+        rtol = 1e-3,
+        atol = 1e-10
+    ),
+    k = SolverSetup(
+        solver      = Bicgstab(), # Bicgstab(), Gmres()
+        preconditioner = Jacobi(),
+        convergence = 1e-7,
+        relax       = 0.7,
+        rtol = 1e-2,
+        atol = 1e-10
+    ),
+    omega = SolverSetup(
+        solver      = Bicgstab(), # Bicgstab(), Gmres()
+        preconditioner = Jacobi(),
+        convergence = 1e-7,
+        relax       = 0.7,
+        rtol = 1e-2,
+        atol = 1e-10
+    ),
+    y = SolverSetup(
+        solver      = Cg(), # Bicgstab(), Gmres()
+        preconditioner = Jacobi(),
+        convergence = 1e-8,
+        rtol = 1e-2,
+        relax       = 0.9,
+    )
+)
+
+runtime = Runtime(iterations=3000, write_interval=100, time_step=1)
+# runtime = Runtime(iterations=2, write_interval=-1, time_step=1)
+
+config = Configuration(
+    solvers=solvers, schemes=schemes, runtime=runtime, hardware=hardware, boundaries=BCs)
+
+
+GC.gc()
+
+initialise!(model.momentum.U, velocity)
+initialise!(model.momentum.p, 0.0)
+initialise!(model.turbulence.k, k_inlet)
+initialise!(model.turbulence.omega, ω_inlet)
+initialise!(model.turbulence.nut, νt_inlet)
+
+residuals = run!(model, config) # 145 iterations
+
+# Reff = stress_tensor(model.momentum.U, nu, model.turbulence.nut)
+# Fp = pressure_force(:wall, model.momentum.p, 1.25)
+# Fv = viscous_force(:wall, model.momentum.U, 1.25, nu, model.turbulence.nut)
+
+
+# plot(; xlims=(0,494))
+# plot!(1:length(Rx), Rx, yscale=:log10, label="Ux")
+# plot!(1:length(Ry), Ry, yscale=:log10, label="Uy")
+# plot!(1:length(Rp), Rp, yscale=:log10, label="p")